phase 3.6: ship semantic router module
